no-op: Move type_syntax to its own package
