+++ resolver/type_syntax/type_syntax.h
