+++ tools/BUILD
@@ -107,6 +107,7 @@ compilation_database(
         "//proto/pay-server:pay-server",
         "//rbi:generate_procs",
         "//resolver:resolver",
+        "//resolver/type_syntax:type_syntax",
         "//rewriter:rewriter",
         "//sorbet_version:sorbet_version",
         "//test:backtrace-test-error",
